services/_archive/services/requirements.txt
blinker==1.4
celery==5.2.1
mailjet-rest==1.3.4
python-dotenv==0.19.2
rollbar==0.16.2
amqp==5.0.7
pymysql==1.0.2
boto3==1.22.2
pusher==3.2.0
Werkzeug==2.0.2